fs: replace FILE.size by f_inode.i_size
A file's size is stored in both the inode and struct filep, so we sync
both values at a number of places.

Instead of duplicating this information, just use the inode's size,
but give it a macro for easier use.

While at it, we also remove redundant assignments to the size field:
The VFS core already does this at truncate time, so repeating it in the
file systems is unnecessary.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250107075939.2841119-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>