fs: retire FILE typdef
The FILE typedef is confusing: ISO C defines the type for buffered I/O,
but we don't have buffered I/O in barebox and instead use it for file
descriptor information, which is struct fd / struct file in Linux.

To remove this confusion, let's retire this type and replace it with
struct file everywhere.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250107075939.2841119-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>