startup: fix infinite loop falling into menu without /env/menu/mainmenu
If MENUFILE (/env/menu/mainmenu) doesn't exist, an infinite loop will
keep executing this non-existent file.

If this happens, let's exit the infinite loop and hang barebox instead.

Previously we dropped to a shell in this situation, but this was changed
as menu-only boots should not be circumventable and drop to a shell due
to errors.

Fixes: 96852a962d0c ("startup: don't fallback to shell from menu")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250107110825.3081026-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>