fs: fix resolving paths when directory name repeats
d_same_name() used to only check if the name exists in any of the
directory's children.

In order to fix lookup of /, the function was changed to also compare
the name against the parent. This is obviously wrong, when the parent
and child have the same name: Resolution of the child would always
return the parent dentry triggering unexpected behavior.

Fix this by effectively reverting the functional change in commit
3a478ff3e1e6 ("fs: support opening /"). This breaks opening /, but that
currently only affects dirfd usage and will be fixed separately.

Fixes: 3a478ff3e1e6 ("fs: support opening /")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250108091425.911629-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>