partitions: dos: allocate NT signature param when recreating partition
We used to only allocate the device parameter when parsing a MBR from
disk, but not when creating the partition dynamically at runtime.

This didn't cause acute problems until we started freeing the parameter
again and triggering a NULL pointer dereference when removing dynamically
created partitions.

Fix this by always allocating the parameter.

Fixes: 19f4033db59e ("partitions: dos: fix memory leaks")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250109112437.3174977-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>