Import ARM Trusted Firmware fiptool utility
On some platforms like the STM32MP, barebox is located in a FIP binary.
For network booting, interacting with the FIP can be quite useful,
so as first step port fiptool(1) to barebox.

Later, we can build a bootm handler or maybe a file system on top.

For now, the command is sufficient to unpack the FIP and manually boot
the artifacts.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250109112501.3175213-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>