scripts: add new scripts_dtc target
Later commit will add a target for building all enabled device trees.
For this to work, we need to build the device tree compiler first, but
there is only a scripts target, which builds many extra tools
unnecessary for a mere device tree compilation.

Therefore, let's split that up and add a new scripts_dtc target that can
be depended on separately.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250112083432.320215-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>