kbuild: allow dependency on any file in images/
The multi defconfigs build a lot of images, which is nice for coverage,
but users may only be interested in the second stage barebox-dt-2nd.img
for example.

To allow them to build only that one without having to wade through
menuconfig, let's allow specifying the image directly, e.g.:

  make images/barebox-dt-2nd.img

And only that image will be built.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250112083432.320215-8-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>