clocksource: timer-ti-dm: add support for K3 SoCs
K3 SoCs have a timer compatible to the ones found on the AM335x SoCs.
This patch adds support for these SoCs to the driver. The main
difference is that on AM335x we do not have clk support and get the
timer frequency from a SoC specific call, thus we have to play some
ifdeffery to add clk support for K3 and still keep the SoC specific
call to the AM335x code.

Link: https://lore.kernel.org/r/20250113-k3-r5-v3-6-065fcdcc28d3@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>