clk: add K3 clk driver
The clk support for K3 SoCs is implemented on the Cortex-R5 boot
processor by the ti-dm firmware binary. The A53 cores access the clks
via mailboxes. However, during early boot we are running barebox on the
Cortex-R5 processor and the ti-dm firmware is not yet running, so we
must implement our own clk driver. Code is based on U-Boot-2025.01-rc1.

Link: https://lore.kernel.org/r/20250113-k3-r5-v3-8-065fcdcc28d3@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>