elf: use iomem regions as fallback when loading to non-sdram memory
The ELF code uses request_sdram_region() to request the regions the
binary shall be copied to. However, not all of these regions are
actually SDRAM. Some specialized ELF files might also use SoC SRAM
which is not registered as SDRAM, so use request_region as a fallback
in these cases.

This is needed on the TI K3 AM625 SoC to successfully load the ti-dm
firmware binary.

Link: https://lore.kernel.org/r/20250113-k3-r5-v3-13-065fcdcc28d3@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>