i.MX8M: HABv4: add option to allow burning the field-return fuse
This adds the required Kconfig options which need to be enabled and
correctly set to build a custom device specific barebox image which can
be used to burn the FIELD_RETURN fuse.

The CST tool can't handle quoted UID strings so we need to define it on
the cmdline by using the -D switch. This removes the quotes within the
CSF file and the CST is happy.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>
Link: https://lore.kernel.org/r/20250123-v2024-05-0-topic-hab-v3-7-e90d0e43c2de@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>