ARM: i.MX93: allow board code to configure the bl33 loadaddr
Add parameterized __imx93_load_and_start_image_via_tfa function to
load_and_start the image on arbitrary load addresses. This should be
taken with caution since it requires to configure the TF-A correctly as
well, therefore the function is prefixt with '__'.

One use case of the new helper is to place OP-TEE at the very beginning
of the DRAM and barebox afterwards.

It has shown that the OPTEE can only operate with the ELE device if
it uses the MU wih messages that are addressed after 0xE0000000.

If OPTEE will be linked to the upper end of the memory layout on systems
with more then 3.5GB the ELE device can therefor not be used anymore.

So moving the layout is crucial for these platforms to properly operate
with OPTEE that is including ELE support.

Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>
Link: https://lore.kernel.org/r/20250128152211.1146024-1-m.grzeschik@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>