i2c: rockchip: Replace driver from kernel repository
The current Rockchip i2c driver was taken from the u-boot repository,
so it has minimal functionality and is designed to be the simplest.
Development with the Rockchip SOC led to the idea that it is not
sufficient to handle some i2c devices (in my case, this i2c EEPROMs
and i2c MUXes).
This patch replaces the driver code from the Linux kernel repository
to make the driver more flexible with SDA/SCL setup/hold settings and
as a result, more stable across boards.
Tested with cusom rk3568-based board.

Signed-off-by: Alexander Shiyan <eagle.alexander923@gmail.com>
Link: https://lore.kernel.org/r/20250204132207.332918-1-eagle.alexander923@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>