partitions: efi: Fix position of alternate gpt when writing
The primary GPT has a pointer to the alternate GPT. Set this pointer
correctly when writing the partition table.

When the GPT originally has been created on the device, the alternate GPT
should be at the end of the device already. However, when the GPT was
created on an image which was then written to the disk, the alternate GPT
might be at the end of the image instead of the device. In this case
the alternate GPT changes its position during a partition write, so we
have to adjust the pointer from the primary to the alternate GPT
accordingly.

Reported-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Link: https://lore.kernel.org/r/20250211093408.1701730-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>