clk: Add clock driver for sun50i-a64
Adds a "one size fit all" clock driver to be used for complex
"composite" clock, composed of: a gate, a mux and severals dividers.
This clk driver could be used to make SoC specific drivers easier
to port from linux sources, but this is not the case yet.

The pll-cpux is set to 816MHz and pll-periph0-2x is set to 1.2GHz.

Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>
Link: https://lore.kernel.org/r/20250212104842.504-3-jmaselbas@zdiv.net
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>