mci: Add sunxi-mmc driver
This driver is adapted from different sources: Linux, u-boot and p-boot.
The latter, p-boot (forked from u-boot), is a bootloader for pinephones.

It currently only support PIO xfer but could be further improved to also
support DMA xfer. This driver is split in three files so it can be used
by the PBL and barebox proper.

Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>
Link: https://lore.kernel.org/r/20250212104842.504-5-jmaselbas@zdiv.net
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>