ARM: k3: add function to detect eMMC boot
Booting from eMMC on K3 SoCs requires some special handling. With SD
card boot we put the environment in the boot FAT partition whereas with
eMMC boot we want to put it into the raw eMMC device. Also with eMMC
boot we have to load the subsequent stages from the eMMC boot partitions
and from the boot FAT otherwise. Add a function to detect eMMC boot
to base these decisions on.

Link: https://lore.kernel.org/r/20250212-k3-emmc-v2-2-8dd1bb0ce60a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>