ARM: k3: add eMMC barebox update handler
This adds a barebox update handler to write a barebox image to the
eMMC boot partitions. The image is written to the currently inactive
boot partition which is made active after a successful write thus
enabling a failsafe update.

The update handler expects an image which has the tiboot3.bin right
at the start of the image and a FIP image containing the next stages
at offset 1MiB. This image can't be generated by barebox itself (as
it requires two barebox images built for different architectures). It
must be generated by an external build system.

Link: https://lore.kernel.org/r/20250212-k3-emmc-v2-14-8dd1bb0ce60a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>