ARM: IMX8MP: var-dart-dt8mcustomboard.dts: use new-style partition binding
The fixed-partitions compatible is meant to be used for the partitions
container holding the individual partitions and not for every individual
partition.

This didn't lead to problems so far, because barebox ignored the
compatible and parsed them as legacy partitions.

Resolve the confusion by changing the description to use the new-style
binding correctly.

While at it, we also drop some leading zeroes for readability's sake.

Cc: Michael Kopfensteiner <michael.kopfensteiner@vahle.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250217180833.3955657-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>