of: partition: add Linux CONFIG_OF_PARTITION-compatible adaptive fixup mode
Linux supports OF partitions for MMC since Linux v6.13-rc1 with commit
2e3a191e89f9 ("block: add support for partition table defined in OF"),
but in a manner incompatible to barebox:

If CONFIG_OF_PARTITION is enabled and a fixed-partitions node exists,
the EFI/MBR partition table is completely ignored.
In barebox, they are allowed to coexist, as long as they don't conflict.

To resolve this conflict, we change the default fixup scheme to an
"adaptive" scheme that uses new-style fixups, but with the compatible
depending on the parent device:

  - If the partitions node in the kernel DT already has a compatible,
    use the same compatible as not to clobber new Linux OF partitions,
    which would be parsed by CONFIG_OF_PARTITION
  - If the device is an MTD device, use "fixed-partitions", so Linux
    sees the same partitions and to maintain backwards compatibility
  - For everything else (i.e. EEPROMs and SD/eMMC), use
    "barebox,fixed-partitions"

This scheme is fully backwards-compatible with older kernels and
userspace:

  - barebox,fixed-partitions replaces fixed-partitions only
    for nodes that Linux didn't parse before anyway

  - fixed-partitions will continue to be emitted for MTD,
    which Linux did parse before

  - barebox-state doesn't actually look for the "fixed-partitions"
    compatible to find out if a patition is in a container.
    Instead it checks the parent node name to be equal
    to "partitions"...

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250217180833.3955657-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>