ARM: boards: Add support for phyBOARD-Segin i.MX 93
This patch adds support for the phyBOARD-Segin i.MX 93 and the phyCORE
i.MX 93 SoM it is based upon.

Note that support is only provided for the standard configuration
featuring 1GB RAM as testing was only possible on that variant and only
the single-configuration lpddr4_timing.c as initially committed to
u-boot in 7c1f8ce3956 worked.

The 1GB configuration from later revisions results in a freeze during
dram setup.

The following functionality has been tested:
 - barebox starts via imxusbloader, tftp and emmc
 - both ethertnet interfaces work
 - barebox_update to emmc
 - boot kernel via network
 - USB OTG
 - USB mass storage on USB Host Port
 - reset

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Link: https://lore.kernel.org/r/20250304-phyboard-segin-imx93-v1-1-8677740a79de@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>