i.MX: HAB: fix i.MX8MP field-return fuse
Current tests showed that fusing just the single field-return bit(0) on
i.MX8MP SoCs brick the device.

All i.MX8M SoCs have a single field-return bit except for the i.MX8MP
which requires a pattern. The pattern is not documented but was
discussed here [1] and the final u-boot commit [2] added the support
accordingly.

This commit is based on the outcome of [1, 2].

[1] https://lore.kernel.org/all/20240621130626.729666-1-paul.geurts@prodrive-technologies.com/
[2] https://github.com/u-boot/u-boot/blob/v2025.01/arch/arm/mach-imx/hab.c#L35

Tested-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>
Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.kernel.org/r/20250310113626.2241630-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>