ARM: boards: remove obsolete PCA9450_RESET_CTRL setup
Since a previous commit, PCA9450_RESET_CTRL is initialized to 0xA1 in
pca9450_probe; explicitly setting this in board.c has no effect anymore.

Remove the setup of this register from the
pca9450_register_init_callback of the two imx93 boards we currently
have.

Configuration of this register in power_init_board() in lowlevel.c of
the imx8 boards remains untouched for now. This is because there, the
register is configured even if the pca9450 driver is not enabled.
Furthermore, the power_init_board() functions initialize a larger array
of pmic registers which I haven't compared against other pmic and
-driver defaults.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Link: https://lore.kernel.org/r/20250313-pca9450-wdog-v2-2-28e97e7da480@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>