firmware: export firmware directory
The recently added k3r5 code needs firmware blobs as part of the image
generation. It hardcodes the firmware path to $(srctree)/firmware which
is the default, but could changed with CONFIG_EXTRA_FIRMWARE_DIR.

Move setting of $(fwdir) to the toplevel Makefile and export it to make
it available to images/Makefile.k3. As fwdir is a bad name for an
exported variable rename it to FIRMWARE_DIR.

Link: https://lore.kernel.org/r/20250314094559.3631184-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>